backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# HTTP client (async)
httpx==0.28.1

# Database
aiosqlite==0.21.0

# Japanese annotation
pykakasi==2.3.0

# LLM API (OpenAI-compatible)
openai>=1.0.0

# Environment config
python-dotenv==1.1.0

# Subtitle parsing
pysubs2==1.7.3

# File upload support
python-multipart==0.0.20

# Encoding detection
chardet==5.2.0